created components in common, moved Loader it
